FastAPI_Server-1/pyproject.toml
[tool.poetry]
name = "api-server"
version = "0.1.0"
description = ""
authors = ["Auro Soni <aurokumar926@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.111.0"
uvicorn = "^0.30.1"
odmantic = "^1.0.2"
motor = "^3.5.0"
starlette-admin = "^0.14.0"
pydantic-settings = "^2.3.4"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.8.0"
celery = "^5.4.0"
flower = "^2.0.1"
redis = "^5.0.8"
authlib = "^1.3.2"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.36"}
asyncpg = "^0.30.0"
psycopg = {extras = ["binary"], version = "^3.2.3"}
alembic = "^1.13.3"
openai = "^1.57.1"
sse-starlette = "^2.1.3"
pandas = "^2.2.3"
wonderwords = "^2.2.0"
better-profanity = "^0.7.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
